rtdetr_paddle/requirements.txt
paddlepaddle-gpu>=2.4.2
tqdm
typeguard
visualdl>=2.2.0
opencv-python <= 4.6.0
PyYAML
shapely
scipy
terminaltables
Cython
pycocotools
setuptools
